+++ lib/Service/SearchService.php
@@ -34,6 +34,8 @@ namespace OCA\FullTextSearch_Solr\Service;
 
 use daita\MySmallPhpTools\Traits\TArrayTools;
 use Exception;
+use OCA\FullTextSearch\Model\SearchRequest;
+use OCA\FullTextSearch_Solr\Utilities\Utils;
 use OCP\FullTextSearch\Model\DocumentAccess;
 use OCP\FullTextSearch\Model\IndexDocument;
 use OCP\FullTextSearch\Model\ISearchResult;
@@ -82,13 +84,30 @@ class SearchService {
         Client $client, ISearchResult $searchResult, DocumentAccess $access
     ) {
 
+        /** @var SearchRequest $request */
+        $request = $searchResult->getRequest();
         $this->logger->debug("Search result requested.",
-                             array("request" => $searchResult->getRequest()->getSearch(),
-                                 "provider" => $searchResult->getProvider()->getId()));
+                             array("request" => $request->getSearch(),
+                                   "provider" => $searchResult->getProvider()->getId()));
 
         $selectQuery = $client->createSelect();
+        $selectQuery->setOmitHeader(false);
 
-        $selectQuery->setQuery($searchResult->getRequest()->getSearch());
+        $selectQuery->setQuery($request->getSearch());
+        $selectQuery->setStart(($request->getPage() -1) * $request->getSize());
+        $selectQuery->setRows($request->getSize());
+
+        // Add the list of tags from the select query to the search request
+        if (!empty($request->getTags())) {
+            $selectQuery->createFilterQuery('tagFilter')->setQuery(Utils::createDocumentField('tags').':'.implode(' ', $request->getTags()));
+        }
+
+        // Add all of the metadata queries to the search request
+        if (!empty($request->getSubTags())) {
+            foreach ($request->getSubTags() as $key => $value) {
+                $selectQuery->createFilterQuery($key)->setQuery(Utils::createDocumentField($key).":{$value}");
+            }
+        }
 
         $resultSet = $client->execute($selectQuery);
 
@@ -98,15 +117,6 @@ class SearchService {
 
     }
 
-
-    /**
-     * @param Client $client
-     * @param string $providerId
-     * @param string $documentId
-     *
-     * @return IndexDocument
-     */
-    //TODO
     /**
      * @param ISearchResult $searchResult
      * @param DocumentAccess $access
@@ -121,7 +131,7 @@ class SearchService {
 
         $searchResult->setTotal($result->response->numFound);
         $searchResult->setMaxScore(intval($result->response->maxScore * 100));  // 100 is arbitrary
-        $searchResult->setTime(1);   // Don't have a means of fetching this value yet. value is milliseconds
+        $searchResult->setTime($result->responseHeader->QTime);
         $searchResult->setTimedOut(false);
 
         foreach ($result->response->docs as $entry) {
@@ -139,7 +149,7 @@ class SearchService {
         $access = new DocumentAccess();
         $access->setViewerId($viewerId);
 
-        list($providerId, $documentId) = explode(':', $entry->id, 2);
+        list($providerId, $documentId) = Utils::parseDocumentIdentifier($entry->id);
         $document = new IndexDocument($providerId, $documentId);
         $document->setAccess($access);
         $document->setHash('Unknown');  // TODO: Save off the hash some where.