+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 
-### v 0.2.0
+## v 0.3.0
+
+- Features
+  - Allow Pagination of results
+  - Index Reset Functionality in place
+  - Update Solarium to 5.0.1
+  - Process Tags and SubTags for queries and search requests
+ 
+## v 0.2.0
 
 Bug Fixes 
 
 - Don't upload folders to Solr
 - Remove the temporary files after executing the extract query.
 
-### v 0.1.0 
+## v 0.1.0 
 
 Initial migration of the code from elasticsearch to solr.